building-agents/sentiment-arbitrage-stock-agent/requirements.txt
flask==3.0.0
yfinance==0.2.36
textblob==0.18.0
python-dotenv==1.0.0
requests==2.31.0
